feat:add auction scheduler

+++ src/main/java/com/example/auction_system_backend/DTO/auction/AuctionResultResponse.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 @Data
 public class AuctionResultResponse {
 
-    private Long itemId;
+    private Integer itemId;
 
     private Long winnerUserId;
